feat: modulos frontend ABD conectados (paginacion 50) + fixes backend verificados con Docker
- Frontend: rutas, unidades, programaciones, incidentes/alertas y reportes
  reemplazan placeholders; PaginadorComponent reutilizable (50/pagina);
  estilos CRUD compartidos; nav y rutas actualizadas
- Backend: endpoints /api/abd/conductores y /api/abd/alertas;
  entidades/DTOs/repositories alineados a integer (esquema V6 real,
  validado por Hibernate contra Docker); ReporteAbdService usa long en conteos
- Generador de datos: corrige numero_disco > varchar(10), placas duplicadas
  por modulo 24, rangos FK derivados de MIN/COUNT (secuencias no son
  transaccionales) y hace idempotentes las tablas de hechos

+++ src/main/java/ec/edu/uteq/sgroas/abd/controller/IncidenteAbdController.java
@@ -32,7 +32,7 @@ public class IncidenteAbdController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> eliminar(@PathVariable Long id) {
+    public ResponseEntity<Void> eliminar(@PathVariable Integer id) {
         incidenteAbdService.eliminar(id);
         return ResponseEntity.noContent().build();
     }
