+++ src/main/java/ec/edu/uteq/sgroas/abd/entity/Ciudad.java
@@ -15,7 +15,7 @@ public class Ciudad {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_ciudad")
-    private Long idCiudad;
+    private Integer idCiudad;
 
     @Column(name = "nombre", nullable = false, length = 100)
     private String nombre;
