+++ src/main/java/ec/edu/uteq/sgroas/abd/entity/Provincia.java
@@ -15,7 +15,7 @@ public class Provincia {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_provincia")
-    private Long idProvincia;
+    private Integer idProvincia;
 
     @Column(name = "nombre", nullable = false, length = 100)
     private String nombre;
