+++ src/main/java/ec/edu/uteq/sgroas/abd/repository/AlertaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlertaRepository extends JpaRepository<Alerta, Long> {
+public interface AlertaRepository extends JpaRepository<Alerta, Integer> {
 
     Page<Alerta> findAllByOrderByIdAlertaDesc(Pageable pageable);
 }
