+++ src/main/java/ec/edu/uteq/sgroas/abd/repository/ConductorAbdRepository.java
@@ -3,5 +3,5 @@ package ec.edu.uteq.sgroas.abd.repository;
 import ec.edu.uteq.sgroas.abd.entity.ConductorAbd;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConductorAbdRepository extends JpaRepository<ConductorAbd, Long> {
+public interface ConductorAbdRepository extends JpaRepository<ConductorAbd, Integer> {
 }
