+++ src/main/java/ec/edu/uteq/sgroas/abd/repository/ProvinciaRepository.java
@@ -3,5 +3,5 @@ package ec.edu.uteq.sgroas.abd.repository;
 import ec.edu.uteq.sgroas.abd.entity.Provincia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProvinciaRepository extends JpaRepository<Provincia, Long> {
+public interface ProvinciaRepository extends JpaRepository<Provincia, Integer> {
 }
