+++ src/main/java/ec/edu/uteq/sgroas/abd/repository/RolAbdRepository.java
@@ -3,5 +3,5 @@ package ec.edu.uteq.sgroas.abd.repository;
 import ec.edu.uteq.sgroas.abd.entity.RolAbd;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolAbdRepository extends JpaRepository<RolAbd, Long> {
+public interface RolAbdRepository extends JpaRepository<RolAbd, Integer> {
 }
