+++ src/main/java/ec/edu/uteq/sgroas/abd/repository/UnidadRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnidadRepository extends JpaRepository<Unidad, Long> {
+public interface UnidadRepository extends JpaRepository<Unidad, Integer> {
 
     Page<Unidad> findByEstadoIgnoreCase(String estado, Pageable pageable);
